feat(appRouter): apply app router to APIs
