+++ src/app/api/auth/[...nextauth]/route.ts
@@ -40,4 +40,5 @@ export const authOptions: NextAuthOptions = {
   },
 };
 
-export default NextAuth(authOptions);
+const handler = NextAuth(authOptions);
+export { handler as GET, handler as POST };
